feat: reorganize CLI commands into shell package
