+++ cmd/polyclient/shell/gui.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later WITH LicenseRef-PolyClient-Plugin-Exception
 
-package commands
+package shell
 
 import (
 	"context"
